step3: 로또 2등

+++ src/main/java/step3/utils/NumberUtils.java
@@ -2,7 +2,7 @@ package step3.utils;
 
 public class NumberUtils {
 
-    public static long getSafeNumber(Long number) {
+    public static long getSafeNumber(Integer number) {
         if (number == null) {
             return 0;
         }
